+++ lib/reactive-ruby/api.rb
