+++ lib/reactive-ruby/callbacks.rb
