+++ lib/reactive-ruby/event.rb
