+++ lib/reactive-ruby/ext/hash.rb
