+++ lib/reactive-ruby/ext/string.rb
