+++ lib/reactive-ruby/isomorphic_helpers.rb
