+++ lib/reactive-ruby/observable.rb
