+++ lib/reactive-ruby/rendering_context.rb
