+++ lib/reactive-ruby/serializers.rb
