+++ lib/reactive-ruby/state.rb
