+++ lib/reactive-ruby/validator.rb
