+++ lib/reactive-ruby/version.rb
